+++ app/views/public/users/unsubscribe.html.erb
@@ -11,10 +11,10 @@
   <div class="row mt-5 pt-5">
     <div class="col-lg-3"></div>
     <div class="col-lg-3 text-center">
-      <%= link_to 'Yse', withdrawal_path(@user.id), method: :patch, "data-confirm" => "本当に退会しますか？", id: "unsubscribe-btn", class: "button" %>
+      <%= link_to 'YES', withdrawal_path(@user.id), method: :patch, "data-confirm" => "本当に退会しますか？", id: "unsubscribe-btn", class: "button" %>
     </div>
     <div class="col-lg-3" style="text-align: center;">
-      <%= link_to 'No', user_path(@user.id), id: "no-btn", class: "button" %>
+      <%= link_to 'NO', user_path(@user.id), id: "no-btn", class: "button" %>
     </div>
     <div class="col-lg-3"></div>
   </div>
